# connector/IoFMT-API.swagger.json
{
  "info": {
    "title": "IoFMT REST API",
    "summary": "This is a REST API for IoFMT Inception project",
    "description": "This REST API acts as a Facade for connecting to SFG20 GraphAPI and maintain a cache to expedite performance.\n        <br><br>\n        This API is also mapped in a Power Platform Custom Connector.\n        <br><br>\n        To generate the API KEY for a customer, please go to: <a href='/admin' target='_blank'>Admin</a>",
    "version": "1.0.0",
    "x-logo": {
      "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"
    }
  },
  "paths": {
...
    "/schedules": {
      "post": {
        "tags": [
          "SFG20"
        ],
        "summary": "Get Schedules",
        "description": "Search SFG20 schedules according to the parameters provided and load into the cache",
        "operationId": "get_schedules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SearchTerm"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/shared-links": {
      "post": {
        "tags": [
          "SFG20"
        ],
        "summary": "Get Shared Links",
        "description": "List the SFG20 shared links available for the user",
        "operationId": "get_shared_links",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigSharedLinks"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/task/complete": {
      "post": {
        "tags": [
          "SFG20"
        ],
        "summary": "Complete Task",
        "description": "Mark a task as completed in SFG20",
        "operationId": "complete_task",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/task_group/complete": {
      "post": {
        "tags": [
          "SFG20"
        ],
        "summary": "Complete Task Group",
        "description": "Mark a group of tasks as completed in SFG20",
        "operationId": "complete_task_group",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/TaskGroup"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/cache": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "Get From Cache",
        "description": "List the data in the cache according to the parameters provided. When a parameter is",
        "operationId": "get_from_cache",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CacheParameters"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      },
      "delete": {
        "tags": [
          "Cache"
        ],
        "summary": "Delete From Cache",
        "description": "Delete all the data in the cache for the selected user",
        "operationId": "delete_from_cache",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/add": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Config Add",
        "description": "Add a new configuration to the Config table",
        "operationId": "config_add",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Config"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/delete/{id}": {
      "delete": {
        "tags": [
          "Config"
        ],
        "summary": "Config Delete",
        "description": "Delete the configuration from the Config table",
        "operationId": "config_delete",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/get/{id}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Config Select",
        "description": "Select a configuration from the Config table",
        "operationId": "config_select",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/token": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Config Select Token",
        "description": "Delete the configuration from the Config table",
        "operationId": "config_select_token",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/shared_links": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Config Shared Links",
        "description": "Get the shared links for the user",
        "operationId": "config_select_shared_links",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ]
      },
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Add Shared Link",
        "description": "Add a new shared link for the user",
        "operationId": "config_add_shared_link",
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SharedLinks"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/config/shared_links/{id}": {
      "delete": {
        "tags": [
          "Config"
        ],
        "summary": "Delete Shared Link",
        "description": "Delete a shared link for the user",
        "operationId": "config_delete_shared_link",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    }
  },
...
  "swagger": "2.0",
  "definitions": {
...
    "ConfigSharedLinks": {
      "properties": {
        "sharelink_id": {
          "type": "string",
          "title": "Sharelink ID",
          "description": "The ID of the SFG20 sharelink"
        },
        "access_token": {
          "type": "string",
          "title": "Access Token",
          "description": "The Access Token provided by SFG20"
        }
      },
      "type": "object",
      "required": [
        "sharelink_id",
        "access_token"
      ],
      "title": "ConfigSharedLinks"
    },
...
    "HTTPValidationError": {
      "properties": {
        "detail": {
          "items": {
            "$ref": "#/definitions/ValidationError"
          },
          "type": "array",
          "title": "Detail"
        }
      },
      "type": "object",
      "title": "HTTPValidationError"
    },
    "Result": {
      "properties": {
        "status": {
          "type": "string",
          "title": "Status",
          "description": "The status of the request"
        },
        "message": {
          "type": "string",
          "title": "Message",
          "description": "The message of the request"
        },
        "data": {
          "items": {
            "type": "object"
          },
          "type": "array",
          "title": "Data",
          "description": "The data of the request"
        }
      },
      "type": "object",
      "required": [
        "status",
        "message",
        "data"
      ],
      "title": "Result"
    },
    "SearchTerm": {
      "properties": {
        "sharelink_id": {
          "type": "string",
          "title": "Sharelink ID",
          "description": "The ID of the SFG20 sharelink"
        },
        "access_token": {
          "type": "string",
          "title": "Access Token",
          "description": "The Access Token provided by SFG20"
        },
        "user_id": {
          "type": "string",
          "title": "User ID",
          "description": "The ID of the user"
        },
        "changes_since": {
          "title": "Changes Since",
          "description": "The date of changes since the last update"
        }
      },
      "type": "object",
      "required": [
        "sharelink_id",
        "access_token",
        "user_id"
      ],
      "title": "SearchTerm"
    },
...
  },
  "securityDefinitions": {
    "APIKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-Access-Token"
    }
  },
...
}